Date input uses AJAX to get term dates from new /calendar/ajax/termdates and update calendar when term is changed.
also added date selector to crossword home page for searching - which doesn't do anything yet

+++ system/application/libraries/Academic_calendar.php
@@ -395,7 +395,7 @@ class Academic_time
 	function AcademicTermWeeks()
 	{
 		$days = self::LengthOfAcademicTerm($this->AcademicYear(), $this->AcademicTerm());
-		return ($days+6) / 7;
+		return (int)(($days+6) / 7);
 	}
 	
 	// TIME
